Update symtable.c
changed int hash to unsigned long hash

+++ symtable.c
@@ -4,8 +4,7 @@ struct Symbol* hashArray[SYMBOL_TABLE_SIZE];
 
 int hash(char *str)
 {
-
-    int hash = 5381;
+    unsigned long hash = 5381;
     int c;
     while ((c = *str++))
         hash = ((hash << 5) + hash) + c;
